Refactoring Hashgeneration für Figuren
    [+] ZobristHasher.java:
            generiert einmalig eine Menge an Pseudozufallszahlen.
            Die Zahlenmenge bleibt immer gleich, da festgelegter Seed.
    [#] Piece.java und Unterklassen:
            geändert, damit ZobristHasher verwendet wird.

+++ src/de/htwsaar/chessbot/engine/model/Piece.java
@@ -62,7 +62,7 @@ public interface Piece {
     *
     * @return ID dieser Figurenart
     */
-    long id();
+    int id();
 
     /**
     * Gib zurück, ob die Figur weiß ist.
